Creating release

See merge request external-unix/inventory!14

+++ image/image.del.php
@@ -21,22 +21,90 @@
     if (check_userlevel($db, $AL_Edit)) {
       logaccess($db, $_SESSION['uid'], $package, "Deleting " . $formVars['id'] . " from images");
 
-      $q_string  = "select img_file ";
+# delete the image from the system
+      $q_string  = "select img_file,img_facing ";
       $q_string .= "from images ";
       $q_string .= "where img_id = " . $formVars['id'];
       $q_images = mysqli_query($db, $q_string) or die($q_string . ": " . mysqli_error($db));
       $a_images = mysqli_fetch_array($q_images);
 
-      if (file_exists($Picturepath . "/" . $a_images['img_file'])) {
-        unlink($Picturepath . "/" . $a_images['img_file']);
+      $q_string  = "select inv_id ";
+      $q_string .= "from inventory ";
+      $q_string .= "where inv_front = " . $formVars['id'] . " or inv_rear = " . $formVars['id'] . " ";
+      $q_inventory = mysqli_query($db, $q_string) or die($q_string . ": " . mysqli_error($db));
+      if (mysqli_num_rows($q_inventory) == 1) {
+        if (file_exists($Picturepath . "/" . $a_images['img_file'])) {
+          unlink($Picturepath . "/" . $a_images['img_file']);
+        }
       }
 
+# now delete it from the database
       $q_string  = "delete ";
       $q_string .= "from images ";
       $q_string .= "where img_id = " . $formVars['id'];
       $insert = mysqli_query($db, $q_string) or die($q_string . ": " . mysqli_error($db));
 
-      print "alert('Image file and data deleted.');\n";
+# find the first front facing one with no file.
+# then find the first  rear facing one with no file.
+# these are the blank or 'unattached' images.
+# if none found, set value to 0.
+# then we loop through the inventory looking for servers that have 
+# the deleted id and replace it with the appropriate unassigned value.
+# An 'unassigned' entry has no img_file value; blank.
+# don't need to query the database twice;
+
+      if ($a_images['img_facing']) {
+
+        $q_string  = "select img_id ";
+        $q_string .= "from images ";
+        $q_string .= "where img_facing = 1 and img_file = '' ";
+        $q_front = mysqli_query($db, $q_string) or die($q_string . ": " . $mysqli_error($db));
+        if (mysqli_num_rows($q_front) > 0) {
+          $a_front = mysqli_fetch_array($q_front);
+        } else {
+          $a_front['img_id'] = 0;
+        }
+
+        $q_string  = "select inv_id ";
+        $q_string .= "from inventory ";
+        $q_string .= "where inv_front = " . $a_front['img_id'] . " ";
+        $q_inventory = mysqli_query($db, $q_string) or die($q_string . ": " . $mysqli_error($db));
+        if (mysqli_num_rows($q_inventory) > 0) {
+          while ($a_inventory = mysql_fetch_array($q_inventory)) {
+            $q_string  = "update ";
+            $q_string .= "inventory ";
+            $q_string .= "set inv_front = " . $a_front['img_id'] . " ";
+            $q_string .= "where inv_id = " . $a_inventory['inv_id'] . " ";
+            $q_result = mysqli_query($db, $q_string) or die($q_string . ": " . $mysqli_error($db));
+          }
+        }
+
+      } else {
+
+        $q_string  = "select img_id ";
+        $q_string .= "from images ";
+        $q_string .= "where img_facing = 0 and img_file = '' ";
+        $q_rear = mysqli_query($db, $q_string) or die($q_string . ": " . $mysqli_error($db));
+        if (mysqli_num_rows($q_rear) > 0) {
+          $a_rear = mysqli_fetch_array($q_rear);
+        } else {
+          $a_rear['img_id'] = 0;
+        }
+
+        $q_string  = "select inv_id ";
+        $q_string .= "from inventory ";
+        $q_string .= "where inv_rear = " . $a_rear['img_id'] . " ";
+        $q_inventory = mysqli_query($db, $q_string) or die($q_string . ": " . $mysqli_error($db));
+        if (mysqli_num_rows($q_inventory) > 0) {
+          while ($a_inventory = mysql_fetch_array($q_inventory)) {
+            $q_string  = "update ";
+            $q_string .= "inventory ";
+            $q_string .= "set inv_rear = " . $a_rear['img_id'] . " ";
+            $q_string .= "where inv_id = " . $a_inventory['inv_id'] . " ";
+            $q_result = mysqli_query($db, $q_string) or die($q_string . ": " . $mysqli_error($db));
+          }
+        }
+      }
 
       print "clear_fields();\n";
     } else {
